@@ -4,8 +4,10 @@
   "description": "A musical notation helper",
   "main": "main.js",
   "scripts": {
-    "start": "electron .",
-    "rebuild": "electron-rebuild -f -w midi"
+    "build": "tsc",
+    "watch": "tsc -w",
+    "lint": "eslint -c .eslintrc --ext .ts ./src",
+    "start": "npm run build && electron ./dist/main.js"
   },
   "keywords": [
     "piano",
@@ -17,19 +19,23 @@
   "author": "B. Jordan",
   "license": "MIT",
   "devDependencies": {
-    "electron": "^4.0.1",
-    "electron-reloader": "^0.3.0"
+    "@types/vexflow": "^3.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.27.0",
+    "@typescript-eslint/parser": "^4.27.0",
+    "electron": "^13.1.6",
+    "eslint": "^7.28.0",
+    "typescript": "^4.3.4"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/blairjordan/pianoboi.git"
   },
   "dependencies": {
-    "electron-rebuild": "^1.8.2",
-    "jquery": "^3.3.1",
+    "@tonaljs/key": "^4.6.0",
+    "@types/jquery": "^3.5.6",
+    "jquery": "^3.6.0",
     "jzz": "^0.8.0",
     "tonal": "^2.1.2",
-    "update-electron-app": "^1.3.0",
     "vexflow": "^1.2.84"
   }
 }
